added i3 config

+++ link.sh
@@ -2,21 +2,25 @@
 set -euxo pipefail
 
 FILES=(
-    bash_aliases            ~/.bash_aliases
-    condarc                 ~/.condarc
-    gdbinit                 ~/.gdbinit
-    profile                 ~/.profile
-    ssh_config              ~/.ssh/config
-    bashrc                  ~/.bashrc
-    config_user-dirs.dirs   ~/.config/user-dirs.dirs
-    taskrc                  ~/.taskrc
+    bashrc          ~/.bashrc
+    condarc         ~/.condarc
+    i3_config       ~/.config/i3/config
+    init.vim        ~/.config/nvim/init.vim
+    snakemake.vim   ~/.config/nvim/snakemake.vim
+    user-dirs       ~/.config/user-dirs.dirs
+    gdbinit         ~/.gdbinit
+    profile         ~/.profile
+    ssh_config      ~/.ssh/config
+    taskrc          ~/.taskrc
+    xinitrc         ~/.xinitrc
 )
 
 mkdir -p ~/trash
 for i in $(seq 0 2 14); do
     src=${FILES[${i}]};
     dst=${FILES[$((i+1))]};
-    mv ${dst} ~/trash;
+    if [[ -e ${dst} ]]; then mv ${dst} ~/trash/${src}; fi;
+    mkdir -p $(dirname ${dst})
     ln -s $(realpath ${src}) ${dst};
 done;
